+++ CHANGELOG.md
@@ -2,13 +2,21 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [36.8.3] 2025-01-30
+## [36.10.1] 2025-01-30
 ### Fixed
 - Warning messages were being incorrectly generated for failing to load a value
   set for an autocomplete field, for which the value set should not be and is
   not loaded at form initialization time.
 - Improved the timing of the formReady event, so that it is sent after the form
   has been rendered.
+
+## [36.10.0] 2025-01-30
+### Added
+- Support for rendering-xhtml on contained[valueSet].expansion.contains.display.
+
+## [36.9.0] 2025-01-30
+### Added
+- Support for rendering-xhtml on answerOption.valueCoding.display.
 
 ## [36.8.2] 2025-01-17
 ### Fixed